some more suggestions
